// Users/Admin/Desktop/react_phone-catalog-solution/react_phone-catalog-solution/src/types/Product.ts
export type Product = {
  id: string;
  category: string;
  namespaceId: string;
  name: string;

  capacityAvailable: string[];
  capacity: string;

  priceRegular: number;
  priceDiscount?: number;

  colorsAvailable: string[];
  color: string;

  images: string[]; 

  description?: {
    title: string;
    text: string[];
  }[];
};